requirements.txt
Flask
Pillow
easyocr
numpy
requests
beautifulsoup4
googlesearch-python
nltk
google-generativeai
python-dotenv